+++ crates/rust-client/src/tests.rs
@@ -28,7 +28,7 @@ use miden_objects::{
     vm::AdviceInputs,
 };
 use miden_tx::utils::{Deserializable, Serializable};
-use rand::rngs::StdRng;
+use rand::{RngCore, rngs::StdRng};
 
 use crate::{
     Client, ClientError,
@@ -41,8 +41,8 @@ use crate::{
     },
 };
 
-async fn insert_new_wallet<R: FeltRng>(
-    client: &mut Client<R>,
+async fn insert_new_wallet(
+    client: &mut Client,
     storage_mode: AccountStorageMode,
     keystore: &FilesystemKeyStore<StdRng>,
 ) -> Result<(Account, Word), ClientError> {
@@ -70,8 +70,8 @@ async fn insert_new_wallet<R: FeltRng>(
     Ok((account, seed))
 }
 
-async fn insert_new_fungible_faucet<R: FeltRng>(
-    client: &mut Client<R>,
+async fn insert_new_fungible_faucet(
+    client: &mut Client,
     storage_mode: AccountStorageMode,
     keystore: &FilesystemKeyStore<StdRng>,
 ) -> Result<(Account, Word), ClientError> {