+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 * Added recency validations for the client (#776).
 * [BREAKING] Changed parameter of `with_expected_output_notes` to accept any type of `OutputNote` (#806).
 * Updated protobuf bindings generation to use `miden-node-proto-build` crate (#807).
+* [BREAKING] Removed `Rng` generic from `Client` and added support for different keystores and RNGs in `ClientBuilder`  (#782).
 
 ### Fixes